requirements.txt
PyQt5~=5.15.10
colorama~=0.4.6
PyYAML~=6.0.1
tqdm~=4.66.2